+++ src/mjlab/scene/scene.py
@@ -15,7 +15,7 @@ from mjlab.sensor import BuiltinSensor, RayCastSensor, Sensor, SensorCfg
 from mjlab.sensor.camera_sensor import CameraSensor
 from mjlab.sensor.sensor_context import SensorContext
 from mjlab.terrains.terrain_entity import TerrainEntity, TerrainEntityCfg
-from mjlab.utils.spec import export_spec
+from mjlab.utils.spec import export_spec, non_default_option_fields
 
 _SCENE_XML = Path(__file__).parent / "scene.xml"
 
@@ -213,6 +213,14 @@ class Scene:
         key_qpos.append(np.array(ent.spec.keys[0].qpos))
         key_ctrl.append(np.array(ent.spec.keys[0].ctrl))
         ent.spec.delete(ent.spec.keys[0])
+      non_default = non_default_option_fields(ent.spec.option)
+      if non_default:
+        fields = ", ".join(non_default)
+        warnings.warn(
+          f"Entity '{ent_name}' has non-default <option> fields ({fields}) that will"
+          " not be propagated by MjSpec.attach(). Use MujocoCfg instead.",
+          stacklevel=2,
+        )
       frame = self._spec.worldbody.add_frame()
       self._spec.attach(ent.spec, prefix=f"{ent_name}/", frame=frame)
     # Add merged keyframe to scene spec.
@@ -233,6 +241,14 @@ class Scene:
     terrain = TerrainEntity(self._cfg.terrain, device=self._device)
     self._terrain = terrain
     self._entities["terrain"] = terrain
+    non_default = non_default_option_fields(terrain.spec.option)
+    if non_default:
+      fields = ", ".join(non_default)
+      warnings.warn(
+        f"Terrain has non-default <option> fields ({fields}) that will not be"
+        " propagated by MjSpec.attach(). Use MujocoCfg instead.",
+        stacklevel=2,
+      )
     frame = self._spec.worldbody.add_frame()
     self._spec.attach(terrain.spec, prefix="", frame=frame)